requirements.txt
openai==0.28
python-dotenv==1.0.1
psycopg2-binary
